// frontend/src/client/models/Annotation.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type Annotation = {
    id: string;
    id_annotation: string;
    id_group: string;
    classe: string;
    family: string;
    genus: string;
    order: string;
    species: string;
    number: number;
    life_stage: string;
    biological_state: string;
    behaviour: string;
    sex: string;
    comments: string;
};